+++ src/zip.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "rapidxml/rapidxml.h"
 #include <string>
+
+#include "rapidxml/rapidxml.h"
 
 std::string zip_buffer(const std::string& zip_path, const std::string& file_path);
 bool zip_has_file(const std::string& zip_path, const std::string& file_path);