+++ NEWS.md
@@ -1,3 +1,5 @@
+# readxl (development version)
+
 # readxl 1.4.5
 
 This release contains no user-facing changes.